+++ packages/datadog_flutter_plugin/lib/src/rum/ddrum.dart
@@ -157,11 +157,17 @@ class DdRum {
 
   /// Notifies that the Exception or Error [error] occurred in currently
   /// presented View, with an origin of [source]. You can optionally set
-  /// additional [attributes] for this error
-  void addError(Object error, RumErrorSource source,
-      {StackTrace? stackTrace, Map<String, Object?> attributes = const {}}) {
+  /// additional [attributes] for this error, an [errorType] and a
+  /// [stackTrace]
+  void addError(
+    Object error,
+    RumErrorSource source, {
+    StackTrace? stackTrace,
+    String? errorType,
+    Map<String, Object?> attributes = const {},
+  }) {
     wrap('rum.addError', logger, attributes, () {
-      return _platform.addError(error, source, stackTrace, {
+      return _platform.addError(error, source, stackTrace, errorType, {
         DatadogPlatformAttributeKey.errorSourceType: 'flutter',
         ...attributes
       });
@@ -170,11 +176,17 @@ class DdRum {
 
   /// Notifies that an error occurred in currently presented View, with the
   /// supplied [message] and with an origin of [source]. You can optionally
-  /// supply a [stackTrace] and send additional [attributes] for this error
-  void addErrorInfo(String message, RumErrorSource source,
-      {StackTrace? stackTrace, Map<String, Object?> attributes = const {}}) {
+  /// supply a [stackTrace], [errorType], and send additional [attributes] for
+  /// this error
+  void addErrorInfo(
+    String message,
+    RumErrorSource source, {
+    StackTrace? stackTrace,
+    String? errorType,
+    Map<String, Object?> attributes = const {},
+  }) {
     wrap('rum.addErrorInfo', logger, attributes, () {
-      return _platform.addErrorInfo(message, source, stackTrace, {
+      return _platform.addErrorInfo(message, source, stackTrace, errorType, {
         DatadogPlatformAttributeKey.errorSourceType: 'flutter',
         ...attributes
       });